[@mantine/core] PinInput: Move to core
