+++ docs/src/docs/core/PinInput.mdx
@@ -1,14 +1,14 @@
 ---
-group: 'mantine-labs'
-package: '@mantine/labs'
+group: 'mantine-core'
+package: '@mantine/core'
 title: PinInput
-slug: /labs/pin-input/
+slug: /core/pin-input/
 category: 'inputs'
 description: 'Capture pin code from the user'
 props: ['PinInput']
-import: "import { PinInput } from '@mantine/labs';"
+import: "import { PinInput } from '@mantine/core';"
 source: 'mantine-labs/src/components/PinInput/PinInput.tsx'
-docs: 'labs/PinInput.mdx'
+docs: 'core/PinInput.mdx'
 styles: ['PinInput']
 ---
