+++ src/mantine-core/src/PinInput/create-pin-array/create-pin-array.ts
