+++ src/mantine-core/src/PinInput/regex/regex.ts
