+++ src/mantine-core/src/TagInput/TagInput.test.tsx
@@ -17,7 +17,7 @@ const defaultProps: TagInputProps = {
   id: 'test-tag-input',
 };
 
-describe('@mantine/labs/TagInput', () => {
+describe('@mantine/core/TagInput', () => {
   checkAccessibility([<TagInput {...defaultProps} />]);
   itSupportsFocusEvents(TagInput, defaultProps, '#test-tag-input');
   itSupportsInputIcon(TagInput, defaultProps);
