+++ src/mantine-demos/src/demos/core/PinInput/PinInput.demos.story.tsx
