+++ src/mantine-demos/src/index.ts
@@ -73,6 +73,7 @@ export * as OverlayDemos from './demos/core/Overlay';
 export * as PaginationDemos from './demos/core/Pagination';
 export * as PaperDemos from './demos/core/Paper';
 export * as PasswordInputDemos from './demos/core/PasswordInput';
+export * as PinInputDemos from './demos/core/PinInput';
 export * as PopoverDemos from './demos/core/Popover';
 export * as ProgressDemos from './demos/core/Progress';
 export * as RadioDemos from './demos/core/Radio';
@@ -117,9 +118,6 @@ export * as NProgressDemos from './demos/nprogress';
 export * as CarouselDemos from './demos/carousel';
 export * as MantineLogoDemos from './demos/logo';
 export * as TipTapDemos from './demos/tiptap';
-
-// @mantine/labs
-export * as PinInputDemos from './demos/labs/PinInput';
 
 // @mantine/dates
 export * as YearPickerDemos from './demos/dates/YearPicker';
