+++ src/mantine-styles-api/src/styles-api/PinInput.styles-api.ts
@@ -1,4 +1,4 @@
-import type { PinInputStylesNames } from '@mantine/labs';
+import type { PinInputStylesNames } from '@mantine/core';
 import { Input } from './Input.styles-api';
 
 export const PinInput: Record<PinInputStylesNames, string> = Input;
